+++ src/snowflake/snowpark/functions.py
@@ -7451,6 +7451,30 @@ def array_to_string(
 
 
 @publicapi
+def array_union_agg(col: ColumnOrName, _emit_ast: bool = True) -> Column:
+    r"""Returns an ARRAY that contains the union of the distinct values from the input
+    arrays in the specified column.
+
+    The values in the returned ARRAY are in no particular order, and the order is not
+    deterministic. The function ignores NULL values in the column and in arrays in the
+    column. If the column contains only NULL values or the input is empty, the function
+    returns an empty ARRAY.
+
+    Args:
+        col: A :class:`Column` object or column name containing arrays with distinct values.
+
+    Example::
+        >>> import re
+        >>> df = session.create_dataframe([[[1, 1, 2]], [[1, 2, 3]]], schema=["a"])
+        >>> row = df.select(array_union_agg("a").alias("result")).collect()[0]
+        >>> sorted(int(i) for i in re.findall(r"\d+", row[0]))
+        [1, 1, 2, 3]
+    """
+    c = _to_col_if_str(col, "array_union_agg")
+    return _call_function("array_union_agg", c, _emit_ast=_emit_ast)
+
+
+@publicapi
 def array_unique_agg(col: ColumnOrName, _emit_ast: bool = True) -> Column:
     """Returns a Column containing the distinct values in the specified column col.
     The values in the Column are in no particular order, and the order is not deterministic.
@@ -11866,7 +11890,7 @@ def regr_slope(y: ColumnOrName, x: ColumnOrName, _emit_ast: bool = True) -> Colu
     Example::
 
         >>> df = session.create_dataframe([[10, 11], [20, 22], [25, None], [30, 35]], schema=["v", "v2"])
-        >>> df = df.group_by("v").agg(regr_slope(df["v2"], df["v"]).alias("regr_slope"))
+        >>> df = df.group_by("v").agg(regr_slope(df["v2"], df["v"]).alias("regr_slope")).order_by("v")
         >>> df.collect()
         [Row(V=10, REGR_SLOPE=None), Row(V=20, REGR_SLOPE=None), Row(V=25, REGR_SLOPE=None), Row(V=30, REGR_SLOPE=None)]
     """