+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Release History
 
+## 1.48.0 (TBD)
+
+### Snowpark Python API Updates
+
+#### Bug Fixes
+
+- Fixed a bug in `Session.client_telemetry` that trace does not have snowflake style trace id.
+
 ## 1.47.0 (TBD)
 
 ### Snowpark Python API Updates
 
 #### New Features
+
+- Added support for the `array_union_agg` function in the `snowflake.snowpark.functions` module.
 
 #### Bug Fixes